Rollback any document changes if adding a nuget package fails.
