+++ src/Features/Core/Portable/AddImport/References/AssemblyReference.cs
