+++ src/Features/Core/Portable/AddImport/References/MetadataSymbolReference.cs
