+++ src/Features/Core/Portable/AddImport/References/ProjectSymbolReference.cs
